Auto refactor init nuts
